Move useFilteredSearchEntityQuery from relation picker to search module
