+++ front/src/modules/companies/components/NewCompanyBoardCard.tsx
@@ -2,8 +2,8 @@ import { useCallback } from 'react';
 
 import { useRecoilScopedState } from '@/recoil-scope/hooks/useRecoilScopedState';
 import { SingleEntitySelect } from '@/relation-picker/components/SingleEntitySelect';
-import { useFilteredSearchEntityQuery } from '@/relation-picker/hooks/useFilteredSearchEntityQuery';
 import { relationPickerSearchFilterScopedState } from '@/relation-picker/states/relationPickerSearchFilterScopedState';
+import { useFilteredSearchEntityQuery } from '@/search/hooks/useFilteredSearchEntityQuery';
 import { getLogoUrlFromDomainName } from '@/utils/utils';
 import { CommentableType, useSearchCompanyQuery } from '~/generated/graphql';
