+++ front/src/modules/search/hooks/useFilteredSearchEntityQuery.ts
